Pull request feedback. Added permanent sentences. Allowed modification of sentences. Allowed bullets to pass through turnstiles.

+++ code/game/machinery/turnstile.dm
@@ -1,20 +1,27 @@
 /obj/machinery/turnstile
 	name = "turnstile"
-	desc = "A remote control switch."
+	desc = "A mechanical door that permits one-way access and prevents tailgating."
 	icon = 'icons/obj/turnstile.dmi'
 	icon_state = "turnstile_map"
 	power_channel = ENVIRON
 	density = TRUE
 	armor = list(melee = 50, bullet = 50, laser = 50, energy = 50, bomb = 10, bio = 100, rad = 100, fire = 90, acid = 70)
 	anchored = TRUE
-	use_power = IDLE_POWER_USE
+	use_power = FALSE
 	idle_power_usage = 2
 	resistance_flags = INDESTRUCTIBLE | LAVA_PROOF | FIRE_PROOF | UNACIDABLE | ACID_PROOF
 	layer = OPEN_DOOR_LAYER
+	var/cuts_off_zipties = FALSE
 
 /obj/machinery/turnstile/Initialize()
 	. = ..()
 	icon_state = "turnstile"
+
+/obj/machinery/turnstile/CanAtmosPass(turf/T)
+	return TRUE
+
+/obj/machinery/turnstile/bullet_act(obj/item/projectile/P, def_zone)
+	return -1 //Pass through!
 
 /obj/machinery/turnstile/CollidedWith(atom/movable/AM)
 	if(ismob(AM))
@@ -39,13 +46,12 @@
 				var/mob/living/carbon/human/HU = AM
 				if(istype(HU))
 					var/obj/item/restraints/handcuffs/cable/zipties = HU.handcuffed
-					if(istype(zipties))
+					if(istype(zipties) && cuts_off_zipties)
 						qdel(zipties)
 						to_chat(usr, "<span class='notice'>\the [src] cuts off \the [zipties].</span>")
 				flick("operate", src)
 				sleep(CONFIG_GET(number/run_delay))
-				AM.x = src.x
-				AM.y = src.y
+				AM.forceMove(loc)
 				playsound(src,'sound/items/ratchet.ogg',50,0,3)
 				var/outdir = null
 				switch(dir)
@@ -59,8 +65,7 @@
 						outdir = EAST
 				var/turf/outturf = get_step(src, outdir)
 				sleep(CONFIG_GET(number/run_delay))
-				AM.x = outturf.x
-				AM.y = outturf.y
+				AM.forceMove(outturf)
 			else
 				flick("deny", src)
 				playsound(src,'sound/machines/deniedbeep.ogg',50,0,3)
